requirements.txt
pyyaml
paramiko
pycrypto
ecdsa
pymongo